% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_levels}
\alias{a_levels}
\title{A-level Attainment Data by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{average_a_level_grade}{Numeric average A-level grade for academic year 2022/23, except for Warrington North (E14001017) and Worthing West (E14001055) which use data from 2023/24}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/find-statistics/a-level-and-other-16-to-18-results/2023-24}{UK Government Explore Education Statistics}
\describe{ Dataset: \code{aggregated_attainment_by_pcon_202024.csv}}
}
\usage{
a_levels
}
\description{
This dataset includes average A-level grades by parliamentary constituency,
}
\keyword{datasets}
