% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{academies}
\alias{academies}
\title{Percentage of academy students by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{weighted_percent_fsm}{Weighted percentage of pupils who attend academies, calculated by dividing total number of academy pupils for the constituency by total number of pupils in the constituency, for academic year 2022/23}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/find-statistics/school-pupils-and-their-characteristics/2023-24}{UK Government Explore Education Statistics}
\describe{Dataset: spc_school_level_underlying_data.csv)}
}
\usage{
academies
}
\description{
This dataset includes the percentage of academy students by parliamentary constituency
}
\keyword{datasets}
