% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_data}
\alias{ei_data}
\title{Joined indicator datasets}
\format{
An object of class \code{data.frame} with 533 rows and 10 columns.
}
\usage{
ei_data
}
\description{
This dataset includes the joined columns for all indicator datasets
}
\keyword{datasets}
