% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_data_standardised}
\alias{ei_data_standardised}
\title{Standardised data}
\format{
An object of class \code{data.frame} with 533 rows and 10 columns.
}
\usage{
ei_data_standardised
}
\description{
This dataset includes the data for all indicators after applying log transformations to funding and percentage of private school students, ensuring consistent directionality by multiplying pupil teacher ratio and percent elligible for free school meals by -1, then standardising using z scores
}
\keyword{datasets}
