% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_index}
\alias{ei_index}
\title{Educational inequality index}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 533 rows and 17 columns.
}
\usage{
ei_index
}
\description{
This dataset includes the educational inequality index scores for all indicators, domains and subdomains for all constituencies, as well as constituency codes and names, and local authority codes and names. Constituencies where missing data was imputed with the mean are marked with asterisks, and constituencies where missing data was imputed with the following year's values are marked with the (^) symbol
}
\keyword{datasets}
