% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_index_complete_case}
\alias{ei_index_complete_case}
\title{Educational inequality index complete case data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 516 rows and 17 columns.
}
\usage{
ei_index_complete_case
}
\description{
This dataset includes the educational inequality index scores for all indicators, domains and subdomains for constituencies with no missing data, as well as constituency codes and names, and local authority codes and names
}
\keyword{datasets}
