% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_index_deciles}
\alias{ei_index_deciles}
\title{Educational inequality index deciles}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 533 rows and 8 columns.
}
\usage{
ei_index_deciles
}
\description{
This dataset includes which deciles constituencies fall under for the domain and subdomain scores
}
\keyword{datasets}
