% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_index_na}
\alias{ei_index_na}
\title{Educational inequality index missing data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 17 columns.
}
\usage{
ei_index_na
}
\description{
This dataset includes the educational inequality index scores for all indicators, domains and subdomains for constituencies with missing data, as well as constituency codes and names (which have been marked with asterisks), and local authority codes and names. Scores were created by imputing the mean for missing values.
}
\keyword{datasets}
