% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ei_la_lookup}
\alias{ei_la_lookup}
\title{Constituency to local authority lookup file}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 533 rows and 4 columns.
}
\source{
\href{https://geoportal.statistics.gov.uk/search?collection=Dataset&q=Parliamentary\%20Constituency\%20to\%20Local\%20Authority\%20District\%20Lookup}{Office for National Statistics}
\describe{Dataset: Ward_to_PCON_to_LAD_to_UTLA_(December_2023)_Lookup_in_the_UK.csv)}
}
\usage{
ei_la_lookup
}
\description{
This dataset includes constituency names and codes with their corresponding local authority names and codes
}
\keyword{datasets}
