% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funding}
\alias{funding}
\title{Pupil funding by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{funding_per_pupil}{Per pupil funding in real terms (not cash terms) for academic year from 2022/23}
}
}
\source{
\href{https://commonslibrary.parliament.uk/research-briefings/cbp-10036/}{UK Parliament House of Commons Library}
\describe{Dataset: Schoolfunding.xlsx (Downloaded from 'school funding' link on page)}
}
\usage{
funding
}
\description{
This dataset includes real values for funding per pupil by parliamentary constituency
}
\keyword{datasets}
