% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcses}
\alias{gcses}
\title{GCSE Attainment Data by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{average_gcse_grade}{Average GCSE attainment 8 score for academic year 2022/23}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/find-statistics/key-stage-4-performance/2022-23}{UK Government Explore Education Statistics}
\describe{Dataset: 2223_sl_pcon_data_revised.csv}
}
\usage{
gcses
}
\description{
This dataset includes average GCSE attainment 8 score by parliamentary constituency
}
\keyword{datasets}
