% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ks2}
\alias{ks2}
\title{KS2 Attainment Data by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{ks2_percent_meeting_standard}{Percent KS2 students meeting expected standard in reading, writing and maths for academic year 2022/23}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/find-statistics/key-stage-2-attainment/2022-23}{UK Government Explore Education Statistics}
\describe{Dataset: ks2_parliamentary_constituency_2019_to_2023_revised.csv}
}
\usage{
ks2
}
\description{
This dataset includes KS2 attainment by parliamentary constituency
}
\keyword{datasets}
