% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progression}
\alias{progression}
\title{Progression to higher education/apprenticeships by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{percent_progressed}{Percent students who progressed to a Level 4 or higher destination post school, which is equivalent to Bachelor's degree or Level 4 apprenticeship, for academic cohort from 2020/21}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/data-catalogue/data-set/8536ca17-09fc-4d90-960e-03c85432c296}{UK Government Explore Education Statistics}
\describe{Dataset: l4_tidy_2023_all_la_final.csv}
}
\usage{
progression
}
\description{
This dataset includes the percent of students who progress to higher education/apprenticeships by parliamentary constituency
}
\keyword{datasets}
