% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_ratio}
\alias{pt_ratio}
\title{Pupil to teacher ratio by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{pupil_to_qual_teacher_ratio}{Pupil to qualified teacher ratio for academic year 2022/23}
}
}
\source{
\href{https://explore-education-statistics.service.gov.uk/data-catalogue/data-set/997087c9-5b1c-4635-b412-83b4bb0130a6}{UK Government Explore Education Statistics}
\describe{Dataset: workforce_ptrs_2010_2023_pcon.csv)}
}
\usage{
pt_ratio
}
\description{
This dataset includes pupil to teacher ratio by parliamentary constituency
}
\keyword{datasets}
