% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_levels}
\alias{a_levels}
\title{A-level Attainment Data by Parliamentary Constituency}
\format{
A data frame with one row per parliamentary constituency:
\describe{
\item{pcon_code}{Parliamentary constituency code}
\item{average_a_level_grade}{Numeric A-level grade score (1–10)}
}
}
\source{
Aggregated from raw data in \code{aggregated_attainment_by_pcon_202024.csv}
}
\usage{
a_levels
}
\description{
This dataset includes average A-level grades by parliamentary constituency,
based on cleaned and imputed Department for Education data.
}
\keyword{datasets}
